renamed "cse-machine" to "ec-evaluator"
